Move labless submit script into labless folder
Add labless/submit_to_labless.py as the maintained contributor submit entrypoint, keep a small root submit.py wrapper for existing users, and document the live labless submission flow in README.md and labless/README.md.
